api/requirements.txt
fastapi
uvicorn
pymongo
supabase
python-dotenv